Fix second-pass Codex review findings for Track F
1. Update stale comment in ast.rs ListIndex to_str_repr: replace
   'Once the F1b parser lands' with accurate description — F1b is
   already merged; the limitation is in the legacy string evaluators

2. Fix 32-bit WASM truncation bug in runtime_resolver.rs ListIndex eval:
   replace `i as usize >= items.len()` with `i >= items.len() as i64`
   to avoid silent truncation on wasm32 where usize is 32-bit

3. Add to_str_repr test for BinOp index expression: documents that
   index expressions inside [...] are never parenthesized (xs[a + b])

+++ crates/goby-wasm/src/runtime_resolver.rs
@@ -672,14 +672,16 @@ impl<'m> RuntimeOutputResolver<'m> {
                 };
                 match list_val {
                     RuntimeValue::ListInt(items) => {
-                        if i < 0 || i as usize >= items.len() {
+                        // Use `i >= items.len() as i64` rather than `i as usize >= items.len()`
+                        // to avoid silent truncation on 32-bit WASM targets where usize is 32-bit.
+                        if i < 0 || i >= items.len() as i64 {
                             self.mark_runtime_abort();
                             return None;
                         }
                         Some(RuntimeValue::Int(items[i as usize]))
                     }
                     RuntimeValue::ListString(items) => {
-                        if i < 0 || i as usize >= items.len() {
+                        if i < 0 || i >= items.len() as i64 {
                             self.mark_runtime_abort();
                             return None;
                         }
